Generics theory and code
