+++ Day 16(3)-Generics/generic_inerface_example/go.mod
@@ -1,0 +1,3 @@
+module numbersOperation
+
+go 1.22.0
