+++ Day 16(3)-Generics/generics_basic_example/go.mod
@@ -1,0 +1,3 @@
+module generics_basic_example
+
+go 1.22.0
