+++ Day 16(3)-Generics/generics_basic_example/main.go
@@ -2,24 +2,24 @@ package main
 
 import "fmt"
 
-func addFloat(a, b float64) float64 {
-	return a + b
+func valueFloat(a float64) float64 {
+	return a
 }
 
-func addInt(a, b int) int {
-	return a + b
+func valueInt(b int) int {
+	return b
 }
 
-func addTypes[T int | float64](a, b T) T {
-	return a + b
+func valueTypes[T any](value T) T {
+	return value
 }
 
 func main() {
-	fmt.Println(addInt(2, 3))
-	fmt.Println(addFloat(2.2, 3.3))
+	fmt.Println(valueInt(2))
+	fmt.Println(valueFloat(3.3))
 
 	//Below is using a generic function
 
-	fmt.Println(addTypes(2, 3))
-	fmt.Println(addTypes(2.2, 3.3))
+	fmt.Println(valueTypes(2))
+	fmt.Println(valueTypes(3.3))
 }
